requirements.txt
meshtastic
paho-mqtt
cryptography
